+++ lib/chef/util/diff.rb
@@ -82,7 +82,7 @@ class Chef
           end
         end
       end
-      
+
       # produces a unified-output-format diff with 3 lines of context
       # ChefFS uses udiff() directly
       def udiff(old_file, new_file)
@@ -185,4 +185,3 @@ class Chef
     end
   end
 end
-